feat(DTFS2-7370): rename unit tests & api test helpers
