+++ trade-finance-manager-api/api-tests/v1/amendments/update-tfm-deal-amendment-completion.api-test.js
@@ -3,7 +3,7 @@ const { generateTfmAuditDetails } = require('@ukef/dtfs2-common/change-stream');
 const { AUDIT_USER_TYPES, AMENDMENT_STATUS } = require('@ukef/dtfs2-common');
 const amendmentController = require('../../../src/v1/controllers/amendment.controller');
 const api = require('../../../src/v1/api');
-const updateFacilityAmendment = require('../utils/updateFacilityAmendment.util');
+const { updateFacilityAmendment } = require('../../helpers/updateFacilityAmendment');
 const MOCK_GEF_AIN_DEAL = require('../../../src/v1/__mocks__/mock-TFM-deal-AIN-submitted');
 const { mockUpdateDeal } = require('../../../src/v1/__mocks__/common-api-mocks');
 const MOCK_USERS = require('../../../src/v1/__mocks__/mock-users');
