--- trade-finance-manager-api/api-tests/v1/deal-builder.js
@@ -1,5 +1,0 @@
-module.exports = (overrides) => ({
-  additionalRefName: 'mock name',
-  bankInternalRefName: 'mock id',
-  ...overrides,
-});
